App code to home page
Code that was in App.js is now in the home page
